% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_search.R
\name{generate_residualsplots}
\alias{generate_residualsplots}
\title{Generate Residual Plots for Model Analysis}
\usage{
generate_residualsplots(
  tab,
  list_pop_param,
  cov_continuous,
  cov_factors,
  result_ML,
  result_5folds,
  i,
  seed = 123
)
}
\arguments{
\item{tab}{A data frame containing the data to be analyzed}

\item{list_pop_param}{A vector of strings indicating the names of population parameters in `tab`.}

\item{cov_continuous}{A vector of strings indicating the names of continuous covariates in `tab`.}

\item{cov_factors}{A vector of strings indicating the names of factor covariates in `tab`.}

\item{result_ML}{A data frame containing model results for maximum likelihood estimation.}

\item{result_5folds}{A data frame containing model results from 5-fold cross-validation.}

\item{i}{An integer indicating the index of the parameter of interest.}

\item{seed}{Numeric value for usage of \code{set.seed()} inside function.}
}
\value{
A list of ggplot objects, each representing a residual plot for a different covariate. The function returns an empty list if no significant relationships are found. It also handles cases where covariates are not selected after the vote.
}
\description{
This function generates residual plots for model analysis based on various covariates and model results. It performs data preprocessing, model fitting using XGBoost, and generates plots to visualize the residuals against the covariates.
}
\examples{
# Assuming 'data' is a data frame with the necessary columns
plots <- generate_residualsplots(data, list_pop_param, cov_continuous, cov_factors, result_ML, result_5folds, 1)

}
