% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_search.R
\name{MLCovSearch}
\alias{MLCovSearch}
\title{MLCovSearch}
\usage{
MLCovSearch(tab, list_pop_param, cov_continuous, cov_factors, seed = NULL)
}
\arguments{
\item{tab}{A data frame containing the data to be analyzed}

\item{list_pop_param}{Parameters}

\item{cov_continuous}{Character vector of continuous covariate names}

\item{cov_factors}{Character vector of categorical or occasion covariate names}
}
\value{
list
}
\description{
MLCovSearch
}
