% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_residuals_plot.R
\name{generate_residuals_plot}
\alias{generate_residuals_plot}
\title{Generate Residual Plots for Model Analysis}
\usage{
generate_residuals_plot(data, result, pop_param, seed = 123)
}
\arguments{
\item{data}{Data frame containing the input variables.}

\item{result}{Results object of class "mlcov_data", obtained from the ml_cov_search function.}

\item{pop_param}{Population parameter name.}

\item{seed}{Numeric value for usage of \code{set.seed()} inside function.}
}
\value{
A list of ggplot objects of length 1 or greater, each representing a residual plot for a different covariate. The function returns an empty list if no significant relationships are found. It also handles cases where covariates are not selected after the vote.
}
\description{
This function generates residual plots for model analysis based on various covariates and model results. It performs data preprocessing, model fitting using XGBoost, and generates plots to visualize the residuals against the covariates.
}
\examples{
# Assuming 'data' is a data frame with the necessary columns
\dontrun{
plots <- generate_residuals_plot(data, 
result, 
pop_param = "V1")
}

}
