% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_shap_summary_plot.R
\name{generate_shap_summary_plot}
\alias{generate_shap_summary_plot}
\title{Generate SHAP Summary Plots}
\usage{
generate_shap_summary_plot(
  data,
  x_bound = NULL,
  dilute = FALSE,
  scientific = FALSE,
  my_format = NULL,
  min_color_bound = "#FFCC33",
  max_color_bound = "#6600CC",
  kind = c("sina", "bar"),
  title = NULL,
  title.position = 0,
  ylab = NULL,
  xlab = NULL
)
}
\arguments{
\item{data}{A list containing required data frames and results.}

\item{x_bound}{use to set horizontal axis limit in the plot}

\item{dilute}{being numeric or logical (TRUE/FALSE), it aims to help make the test
plot for large amount of data faster. If dilute = 5 will plot 1/5 of the
data. If dilute = TRUE or a number, will plot at most half points per
feature, so the plotting won't be too slow. If you put dilute too high, at
least 10 points per feature would be kept. If the dataset is too small
after dilution, will just plot all the data}

\item{scientific}{show the mean|SHAP| in scientific format. If TRUE, label
format is 0.0E-0, default to FALSE, and the format will be 0.000}

\item{my_format}{supply your own number format if you really want}

\item{min_color_bound}{min color hex code for colormap. Color gradient is
scaled between min_color_bound and max_color_bound. Default is "#FFCC33".}

\item{max_color_bound}{max color hex code for colormap. Color gradient is
scaled between min_color_bound and max_color_bound. Default is "#6600CC".}

\item{kind}{By default, a "sina" plot is shown. As an alternative,
set \code{kind = "bar"} to visualize mean absolute SHAP values as a
barplot. Its color is controlled by \code{max_color_bound}. Other
arguments are ignored for this kind of plot.}

\item{title}{A character string to customize the title}

\item{title.position}{A numeric value from 0.0-1.0 to adjust the alignment of the title}

\item{ylab}{A character string to customize the y-axis}

\item{xlab}{A character string to customize the x-axis}
}
\value{
A list of ggplot objects, each representing a SHAP summary plot for a different parameter.
}
\description{
This function generates SHAP summary plots for the XGBoost model.
}
\examples{
# Assuming 'data' is a list with necessary components
\dontrun{
plots <- generate_shap_summary_plot(data, title = "Custom Title", y.inter = 0.25, ...)
}

}
