% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_search.R
\name{ml_cov_search}
\alias{ml_cov_search}
\title{ml_cov_search}
\usage{
ml_cov_search(data, pop_param, cov_continuous, cov_factors, seed = 123)
}
\arguments{
\item{data}{A data frame containing the data to be analyzed}

\item{pop_param}{Character vector of population parameter names}

\item{cov_continuous}{Character vector of continuous covariate names}

\item{cov_factors}{Character vector of categorical or occasion covariate names}

\item{seed}{Numeric value for usage of \code{set.seed()} inside function.}
}
\value{
list
}
\description{
ml_cov_search
}
