% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mlcov_data.R
\name{print.mlcov_data}
\alias{print.mlcov_data}
\title{Print \code{mlcov_data}}
\usage{
\method{print}{mlcov_data}(x, ...)
}
\arguments{
\item{x}{An \code{mlcov_data}.}

\item{...}{Additional args.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print generic used to return information about \code{mlcov_data} object
}
